Refactor tests for Yoast PHP Unit Polyfills
