+++ dev/composer.json
@@ -2,17 +2,35 @@
   "prefer-stable": true,
   "require": {
     "ext-json": "*",
-    "php": ">=7.0",
-    "yoast/phpunit-polyfills": "2.0.0"
+    "php": ">=7.0"
   },
   "config": {
     "optimize-autoloader": true,
     "preferred-install": "dist",
     "autoload": {
       "classmap": [
-        "./../src/",
-        "./../tests/"
+        "../src/",
+        "../tests/"
+      ],
+      "psr-4": {
+        "WordPress\\": "../src/WordPress",
+        "Pimple\\": "../src/pimple/pimple/src/Pimple",
+        "Psr\\Container\\": "../src/psr/container",
+        "Psr\\SimpleCache\\": "../src/psr/simple-cache",
+        "Symfony\\Component\\Process\\": "../src/Symfony/process",
+        "Symfony\\Component\\EventDispatcher\\": "../src/Symfony/event-dispatcher",
+        "Symfony\\Component\\Filesystem\\": "../src/Symfony/filesystem",
+        "Opis\\JsonSchema\\": "../src/opis/json-schema/src",
+        "Opis\\String\\": "../src/opis/string/src",
+        "Opis\\Uri\\": "../src/opis/uri/src"
+      },
+      "files": [
+        "../src/WordPress/Blueprints/functions.php",
+        "../src/WordPress/Zip/functions.php",
+        "../src/WordPress/Streams/stream_str_replace.php",
+        "../src/WordPress/AsyncHttp/async_http_streams.php"
       ]
     }
   }
 }
+
