+++ phpunit.xml.dist
@@ -7,7 +7,7 @@
 >
     <testsuites>
         <testsuite name="default">
-            <directory>./../tests</directory>
+            <directory>tests</directory>
         </testsuite>
     </testsuites>
 </phpunit>
