+++ tests/unit/configuration/PhpUnitTestCase.php
@@ -1,0 +1,5 @@
+<?php
+
+use Yoast\PHPUnitPolyfills\TestCases\TestCase;
+
+abstract class PHPUnitTestCase extends TestCase {}
