+++ tests/unit/json_mapper/resources/Bag.php
