+++ tests/unit/json_mapper/resources/TestResourceClassComplexMapping.php
