+++ tests/unit/json_mapper/resources/TestResourceClassSetValue.php
